+++ src/frontend/org/voltdb/utils/VoltTypeUtil.java
@@ -358,24 +358,4 @@ public abstract class VoltTypeUtil {
         }
         return vt;
     }
-
-    public static enum TABLE_TYPE {
-        PERSISTENT(0),              // Regular PersistentTable
-        STREAM_VIEW_ONLY(1),        // StreamTable without ExportTupleStream (Views only)
-        STREAM(2),                  // StreamTable with ExportTupleStream
-        PERSISTENT_MIGRATE(3),      // PersistentTable with associated Stream for migrating DELETES
-        PERSISTENT_EXPORT(4);       // PersistentTable with associated Stream for leaking INSERTS
-
-        final int type;
-        TABLE_TYPE(int type) {
-            this.type = type;
-        }
-        public int get() {
-            return type;
-        }
-
-        public static boolean isStream(int e) {
-            return (e ==STREAM.get() || e == STREAM_VIEW_ONLY.get());
-        }
-    }
 }
