+++ tests/frontend/org/voltdb/catalog/TestCatalogDiffs.java
@@ -759,7 +759,7 @@ public class TestCatalogDiffs extends TestCase {
         builder.addLiteralSchema("\nALTER TABLE A USING TTL 20 MINUTES ON COLUMN C2 MAX_FREQUENCY 3 BATCH_SIZE 10;");
         assertTrue("Failed to compile schema", builder.compile(testDir + File.separator + "testAlterTableTTL22.jar"));
         Catalog catUpdated2 = catalogForJar(testDir + File.separator + "testAlterTableTTL22.jar");
-       verifyDiff(catUpdated1, catUpdated2, false, null, true, false, true);
+        verifyDiff(catUpdated1, catUpdated2, false, null, true, false, true);
 
         builder.addLiteralSchema("\nALTER TABLE A DROP TTL;");
         assertTrue("Failed to compile schema", builder.compile(testDir + File.separator + "testAlterTableTTL3.jar"));
